Adapt radio button question design to ereferrals.
