+++ app/src/ereferrals/res/layout/multi_question_radio_button_row.xml
@@ -8,6 +8,11 @@
     android:gravity="center_vertical"
     android:orientation="vertical">
 
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="0.5dp"
+        android:background="@color/light_grey_purple" />
+
     <LinearLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -15,7 +20,6 @@
         android:paddingLeft="@dimen/question_padding"
         android:paddingRight="@dimen/question_padding">
 
-        <include layout="@layout/bottom_view" />
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -31,56 +35,47 @@
                 android:gravity="left"
                 android:paddingBottom="@dimen/image_padding"
                 android:paddingTop="@dimen/image_padding"
-                android:scaleType="fitCenter" />
+                android:scaleType="fitCenter"
+                android:visibility="gone" />
 
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical">
 
-                <LinearLayout
+            <RelativeLayout
                     android:layout_width="match_parent"
                     android:layout_height="match_parent"
                     android:layout_gravity="right"
                     android:layout_marginBottom="@dimen/input_edit_text_top_bottom"
                     android:layout_marginLeft="@dimen/input_edit_text_margin"
-                    android:layout_marginRight="@dimen/input_edit_text_margin"
-                    android:layout_marginTop="@dimen/input_edit_text_top_bottom"
-                    android:gravity="right"
-                    android:orientation="horizontal"
-                    android:weightSum="1">
+                android:layout_marginTop="@dimen/input_edit_text_top_bottom">
 
                     <org.eyeseetea.sdk.presentation.views.CustomTextView
                         android:id="@+id/row_header_text"
-                        android:layout_width="0dp"
+                        android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
-                        android:layout_gravity="left"
-                        android:layout_weight="0.9"
                         android:background="@color/transparent"
                         android:gravity="left"
                         android:paddingLeft="@dimen/input_edit_text_left_padding"
                         android:paddingRight="@dimen/input_edit_text_right_padding"
-                        android:textColor="@color/dark_purple"
-                        android:textSize="?attr/font_small"
+                        android:textColor="@color/text_first_color"
+                        android:textColorHint="@color/text_first_color"
+                        android:textSize="?attr/font_xsmall"
                         app:font_name="@string/normal_font" />
 
-                    <ImageView
+                <include
+                    layout="@layout/radio"/>
+
+                <ImageView
                         android:id="@+id/row_header_compulsory"
-                        android:layout_width="0dp"
+                    android:layout_width="10dp"
                         android:layout_height="10dp"
-                        android:layout_gravity="right|center_vertical"
-                        android:layout_toRightOf="@+id/row_header_text"
-                        android:layout_weight="0.1"
-                        android:gravity="right|center_vertical"
+                    android:layout_gravity="center_vertical"
+                    android:gravity="center_vertical"
                         android:scaleType="fitCenter"
                         android:src="@drawable/compulsory"
-                        android:visibility="invisible" />
-                </LinearLayout>
+                    android:visibility="gone" />
+            </RelativeLayout>
 
-                <include layout="@layout/radio" />
 
             </LinearLayout>
-        </LinearLayout>
 
     </LinearLayout>
 </LinearLayout>
