+++ app/src/ereferrals/res/layout/uncheckeable_radiobutton.xml
@@ -19,18 +19,10 @@
 
 <org.eyeseetea.sdk.presentation.views.CustomRadioButton xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    style="@style/RadioButtonNullStyle"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    android:layout_gravity="right|center_vertical"
-    android:layout_margin="0dp"
-    android:drawablePadding="@dimen/input_edit_text_left_padding"
-    android:drawableRight="@drawable/radiobutton"
-    android:gravity="right|center_vertical"
-    android:paddingBottom="2dip"
-    android:paddingLeft="@dimen/input_edit_text_left_padding"
-    android:paddingRight="@dimen/input_edit_text_left_padding"
-    android:paddingTop="2dip"
+    android:layout_gravity="center_vertical"
+    android:gravity="center_vertical"
     android:text=""
     android:textSize="?attr/font_small"
     app:font_name="@string/normal_font" />
