+++ app/src/laos/java/org/eyeseetea/malariacare/views/question/multiquestion/strategies/RadioButtonMultiquestionViewStrategy.java
@@ -1,0 +1,10 @@
+package org.eyeseetea.malariacare.views.question.multiquestion.strategies;
+
+import org.eyeseetea.malariacare.views.question.multiquestion.RadioButtonMultiQuestionView;
+
+public class RadioButtonMultiquestionViewStrategy extends ARadioButtonMultiquestionViewStrategy {
+    public RadioButtonMultiquestionViewStrategy(
+            RadioButtonMultiQuestionView radioButtonMultiQuestionView) {
+        super(radioButtonMultiQuestionView);
+    }
+}
